+++ src/RestApi/BasketRestController.php
@@ -31,20 +31,10 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
 final class BasketRestController extends AbstractFOSRestController
 {
     // literal constants
-    private const TIME_TS = 'time_ts';
-    private const DESCRIPTION = 'description';
-    private const PICTURE = 'picture';
-    private const UPDATED_AT = 'updatedAt';
     private const STATUS = 'status';
-    private const CONTACT_TYPES = 'contactTypes';
-    private const MOBILE_NUMBER = 'handy';
     private const NOT_LOGGED_IN = 'not logged in';
-    private const ID = 'id';
-    private const CREATED_AT = 'createdAt';
-    private const REQUESTS = 'requests';
     private const LAT = 'lat';
     private const LON = 'lon';
-    private const TEL = 'tel';
     private const MAX_BASKET_DISTANCE = 50;
 
     public function __construct(
@@ -56,33 +46,24 @@ final class BasketRestController extends AbstractFOSRestController
     ) {
     }
 
+    // TODO rework api description
     /**
-     * Returns a list of baskets depending on the type.
-     * 'mine': lists all baskets of the current user.
-     * 'coordinates': lists all basket IDs together with the coordinates.
+     * Returns all current baskets of the user.
      *
      * Returns 200 and a list of baskets or 401 if not logged in.
      *
      * @OA\Tag(name="basket")
      */
-    #[Rest\Get('baskets')]
-    #[Rest\QueryParam(name: 'type', requirements: '(mine|coordinates)', default: 'mine')]
-    public function listBaskets(ParamFetcher $paramFetcher): Response
+    #[Rest\Get('user/current/baskets')]
+    public function listBaskets(): Response
     {
         $baskets = [];
-        switch ($paramFetcher->get('type')) {
-            case 'mine':
-                if (!$this->session->mayRole()) {
-                    throw new UnauthorizedHttpException('', self::NOT_LOGGED_IN);
-                }
-                $baskets = $this->getCurrentUsersBaskets();
-                break;
-            case 'coordinates':
-                $baskets = $this->gateway->getBasketCoordinates();
-                break;
+        if (!$this->session->mayRole()) {
+            throw new UnauthorizedHttpException('', self::NOT_LOGGED_IN);
         }
+        $baskets = $this->basketTransactions->getCurrentUsersBaskets();
 
-        return $this->handleView($this->view(['baskets' => $baskets], 200));
+        return $this->handleView($this->view($baskets, 200));
     }
 
     /**
@@ -111,68 +92,8 @@ final class BasketRestController extends AbstractFOSRestController
         }
 
         $baskets = $this->gateway->listNearbyBasketsByDistance($this->session->id(), $location, $distance);
-        $baskets = array_map(function ($b) {
-            $basket = $this->gateway->getBasket((int)$b[self::ID]);
-            $request = $this->gateway->getRequest($basket[self::ID], $this->session->id(), $basket['foodsaver_id']);
-            if ($request) {
-                $request = [$request];
-            }
 
-            return $this->normalizeBasket($basket, $request);
-        }, $baskets);
-
-        return $this->handleView($this->view(['baskets' => $baskets], 200));
-    }
-
-    private function getCurrentUsersBaskets(): array
-    {
-        $updates = $this->gateway->listUpdates($this->session->id());
-        $baskets = $this->gateway->listMyBaskets($this->session->id());
-        $baskets = array_map(fn ($b) => $this->normalizeMyBasket($b, $updates), $baskets);
-
-        return $baskets;
-    }
-
-    /**
-     * Normalizes the details of a basket of the current user for the Rest
-     * response, including requests.
-     *
-     * @param array $basketData basket data
-     * @param array $updates list of updates
-     */
-    private function normalizeMyBasket(array $basketData, array $updates = []): array
-    {
-        $basket = [
-            self::ID => (int)$basketData[self::ID],
-            self::DESCRIPTION => html_entity_decode((string)$basketData[self::DESCRIPTION]),
-            self::PICTURE => $basketData[self::PICTURE],
-            self::CREATED_AT => (int)$basketData[self::TIME_TS],
-            self::UPDATED_AT => (int)$basketData[self::TIME_TS],
-            self::REQUESTS => []
-        ];
-
-        // add requests, if there are any in the updates
-        foreach ($updates as $update) {
-            if ((int)$update[self::ID] === $basket[self::ID]) {
-                $basket[self::REQUESTS][] = $this->normalizeRequest($update);
-                $basket[self::UPDATED_AT] = max($basket[self::UPDATED_AT], (int)$update[self::TIME_TS]);
-            }
-        }
-
-        return $basket;
-    }
-
-    /**
-     * Normalizes a basket request.
-     */
-    private function normalizeRequest(array $request): array
-    {
-        $user = RestNormalization::normalizeUser($request, 'fs_');
-
-        return [
-            'user' => $user,
-            'time' => $request[self::TIME_TS],
-        ];
+        return $this->handleView($this->view($baskets, 200));
     }
 
     /**
@@ -189,65 +110,10 @@ final class BasketRestController extends AbstractFOSRestController
         }
 
         $basket = $this->gateway->getBasket($basketId);
+
         $this->verifyBasketIsAvailable($basket);
-        if ($basket['fs_id'] == $this->session->id()) {
-            $requests = $this->gateway->listRequests($basketId, $this->session->id());
-        } else {
-            $requests = $this->gateway->getRequest($basketId, $this->session->id(), $basket['foodsaver_id']);
-            if ($requests) {
-                $requests = [$requests];
-            }
-        }
-        $basket = $this->normalizeBasket($basket, $requests);
 
-        return $this->handleView($this->view(['basket' => $basket], 200));
-    }
-
-    /**
-     * Normalizes the details of a basket for the Rest response.
-     *
-     * @param array $basketData the basket data
-     */
-    private function normalizeBasket(array $basketData, array $updates = []): array
-    {
-        // set main properties
-        $creator = RestNormalization::normalizeUser($basketData, 'fs_');
-        $basket = [
-            self::ID => (int)$basketData[self::ID],
-            self::STATUS => (int)$basketData[self::STATUS],
-            self::DESCRIPTION => html_entity_decode((string)$basketData[self::DESCRIPTION]),
-            self::PICTURE => $basketData[self::PICTURE],
-            self::CONTACT_TYPES => array_map('\intval', explode(':', (string)$basketData['contact_type'])),
-            self::CREATED_AT => (int)$basketData[self::TIME_TS],
-            self::UPDATED_AT => (int)$basketData[self::TIME_TS],
-            'until' => (int)$basketData['until_ts'],
-            self::LAT => (float)$basketData[self::LAT],
-            self::LON => (float)$basketData[self::LON],
-            'creator' => $creator,
-            'requestCount' => $basketData['request_count'],
-            self::REQUESTS => []
-        ];
-
-        // add phone numbers if contact_type includes telephone
-        $tel = '';
-        $handy = '';
-        $telephoneContactType = 2;
-        if (isset($basketData['contact_type']) && \in_array($telephoneContactType, $basket[self::CONTACT_TYPES], true)) {
-            $tel = $basketData[self::TEL];
-            $handy = $basketData[self::MOBILE_NUMBER];
-        }
-        $basket[self::TEL] = $tel;
-        $basket[self::MOBILE_NUMBER] = $handy;
-
-        // add requests, if there are any in the updates
-        foreach ($updates as $update) {
-            if ((int)$update[self::ID] === $basket[self::ID]) {
-                $basket[self::REQUESTS][] = $this->normalizeRequest($update);
-                $basket[self::UPDATED_AT] = max($basket[self::UPDATED_AT], (int)$update[self::TIME_TS]);
-            }
-        }
-
-        return $basket;
+        return $this->handleView($this->view($basket, 200));
     }
 
     /**
@@ -277,16 +143,6 @@ final class BasketRestController extends AbstractFOSRestController
         }
 
         return $this->getBasket($basketId);
-    }
-
-    /**
-     * Checks if the number is a valid value in the given range.
-     * TODO Duplicated in FoodSharePointRestController.php.
-     */
-    private function isValidNumber($value, float $lowerBound, float $upperBound): bool
-    {
-        return !is_null($value) && !is_nan($value)
-            && ($lowerBound <= $value) && ($upperBound >= $value);
     }
 
     /**
@@ -335,15 +191,19 @@ final class BasketRestController extends AbstractFOSRestController
         if (!$this->session->mayRole()) {
             throw new UnauthorizedHttpException('', self::NOT_LOGGED_IN);
         }
+        $existingBasket = $this->gateway->getBasket($basketId);
 
-        $this->findEditableBasket($basketId);
+        $this->verifyBasketIsAvailable($existingBasket);
+        if ($existingBasket->creator->id !== $this->session->id()) {
+            throw new UnauthorizedHttpException('', 'You are not the owner of the basket.');
+        }
+
         $errors = $validator->validate($basket);
         if ($errors->count() > 0) {
             $firstError = $errors->get(0);
             throw new BadRequestHttpException(json_encode(['field' => $firstError->getPropertyPath(), 'message' => $firstError->getMessage()]));
         }
 
-        //update basket
         $this->basketTransactions->editBasket($basketId, $basket);
 
         return $this->getBasket($basketId);
@@ -373,7 +233,7 @@ final class BasketRestController extends AbstractFOSRestController
         $basket = $this->gateway->getBasket($basketId);
         $this->verifyBasketIsAvailable($basket);
 
-        $basketCreatorId = $basket['foodsaver_id'];
+        $basketCreatorId = $basket->creator->id;
 
         // check for existing request
         $requestStatus = $this->gateway->getRequestStatus($basketId, $this->session->id(), $basketCreatorId);
@@ -417,40 +277,20 @@ final class BasketRestController extends AbstractFOSRestController
     }
 
     /**
-     * Finds and returns the user's basket with the given id. Throws HttpExceptions
-     * if the basket does not exist, was deleted, or is owned by a different user.
-     *
-     * @param int $basketId id of a basket
-     *
-     * @return array the basket's entry from the database
-     */
-    private function findEditableBasket(int $basketId): array
-    {
-        $basket = $this->gateway->getBasket($basketId);
-
-        $this->verifyBasketIsAvailable($basket);
-        if ($basket['fs_id'] !== $this->session->id()) {
-            throw new UnauthorizedHttpException('', 'You are not the owner of the basket.');
-        }
-
-        return $basket;
-    }
-
-    /**
      * Verifies that the basket was not deleted and is not expired. Otherwise this
      * method throws an appropriate HttpException.
      */
-    private function verifyBasketIsAvailable(array $basket): void
+    private function verifyBasketIsAvailable(?Basket $basket): void
     {
-        if (!$basket || $basket[self::STATUS] === BasketStatus::DELETED_OTHER_REASON) {
+        if (!$basket || $basket->status === BasketStatus::DELETED_OTHER_REASON) {
             throw new NotFoundHttpException('Basket does not exist.');
         }
 
-        if ($basket[self::STATUS] === BasketStatus::DELETED_PICKED_UP) {
+        if ($basket->status === BasketStatus::DELETED_PICKED_UP) {
             throw new NotFoundHttpException('Basket was already picked up.');
         }
 
-        if ($basket['until_ts'] < time()) {
+        if ($basket->until < time()) {
             throw new NotFoundHttpException('Basket is expired.');
         }
     }
@@ -476,17 +316,26 @@ final class BasketRestController extends AbstractFOSRestController
         if (!$this->isValidNumber($lat, -90.0, 90.0) || !$this->isValidNumber($lon, -180.0, 180.0)) {
             if ($defaultLocation !== null) {
                 return $defaultLocation;
-            } else {
-                // find user's location
-                $loc = $this->session->user('location');
-                if (!$loc || ($loc->lat === 0 && $loc->lon === 0)) {
-                    throw new BadRequestHttpException('The user profile has no address.');
-                }
-                $lat = $loc->lat;
-                $lon = $loc->lon;
             }
+            // find user's location
+            $loc = $this->session->user('location');
+            if (!$loc || ($loc->lat === 0 && $loc->lon === 0)) {
+                throw new BadRequestHttpException('The user profile has no address.');
+            }
+            $lat = $loc->lat;
+            $lon = $loc->lon;
         }
 
         return GeoLocation::createFromArray(['lat' => $lat, 'lon' => $lon]);
+    }
+
+    /**
+     * Checks if the number is a valid value in the given range.
+     * TODO Duplicated in FoodSharePointRestController.php.
+     */
+    private function isValidNumber($value, float $lowerBound, float $upperBound): bool
+    {
+        return !is_null($value) && !is_nan($value)
+            && ($lowerBound <= $value) && ($upperBound >= $value);
     }
 }